first candidate

+++ laravel/resources/views/admin/index/news.blade.php
@@ -31,7 +31,7 @@
           <td>
             <a class="btn btn-success" href="{{ '/news/' .$n->id . '-' . sluggify($n->title) }}"><i class="fa fa-eye"></i></a>
             <a class="btn btn-primary" href="/admin/news/edit/{{ $n->id }}"><i class="fa fa-pencil"></i></a>
-            <a class="btn btn-danger" href="/admin/news/destroy/{{ $n->id }}"><i class="fa fa-trash"></i></a>
+            <a class="btn btn-danger" data-confirm="la news {{ $n->title }}" href="/admin/news/destroy/{{ $n->id }}"><i class="fa fa-trash"></i></a>
           </td>
         </tr>
         @endforeach
@@ -40,4 +40,5 @@
   </div>
 
 </section>
+
 @stop
