+++ laravel/resources/views/admin/index/page.blade.php
@@ -26,7 +26,7 @@
           <td>
             <a class="btn btn-success" href="/page/{{ $page->id . '-' . sluggify($page->title) }}"><i class="fa fa-eye"></i></a>
             <a class="btn btn-primary" href="/admin/page/edit/{{ $page->id }}"><i class="fa fa-pencil"></i></a>
-            <a class="btn btn-danger" href="/admin/page/destroy/{{ $page->id }}"><i class="fa fa-trash"></i></a>
+            <a class="btn btn-danger" data-confirm="la page {{ $page->title }}" href="/admin/page/destroy/{{ $page->id }}"><i class="fa fa-trash"></i></a>
           </td>
         </tr>
         @endforeach
@@ -35,4 +35,5 @@
   </div>
 
 </section>
+
 @stop
