+++ laravel/resources/views/includes/navbar.blade.php
@@ -11,28 +11,27 @@
     </div>
     <div id="navbar" class="navbar-collapse collapse">
       <ul class="nav navbar-nav">
-        <li class="active"><a href="/news/">News</a></li>
+        <li class="{{ Request::is('news') ? 'active' : '' }}"><a href="/news/">News</a></li>
         <li class="dropdown">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Informations <span class="caret"></span></a>
           <ul class="dropdown-menu">
             <li><a href="/about"><i class="fa fa-fw fa-file-pdf-o"></i> Présentation du parcours</a></li>
-
+            <li><a href="/important-calendar"><span class="glyphicon glyphicon glyphicon-calendar"></span> Dates importantes</a></li>
           </ul>
         </li>
       </ul>
       <ul class="nav navbar-nav navbar-right">
         @if(Auth::check())
-        <li class="dropdown">
+        <li class="dropdown {{ Request::is('calendar') || Request::is('profile') ? 'active' : '' }}">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">{{ Auth::user()->first_name }} <span class="caret"></span></a>
           <ul class="dropdown-menu">
             <li><a href="/calendar"><span class="glyphicon glyphicon glyphicon-calendar"></span> Mon planning</a></li>
-            <li><a href="/profile"><span class="glyphicon glyphicon glyphicon-user"></span> Gérer mon profil (WIP)</a></li>
             <li role="separator" class="divider"></li>
             <li><a href="/logout"><span class="glyphicon glyphicon glyphicon-log-out"></span> Se déconnecter</a></li>
           </ul>
         </li>
         @if(Auth::user()->isAdmin())
-        <li class="dropdown">
+        <li class="dropdown {{ Request::is('admin/*') ? 'active' : '' }}">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Administration <span class="caret"></span></a>
           <ul class="dropdown-menu">
             <li><a href="/admin/news/create"><i class="fa fa-fw fa-pencil-square-o"></i> Ecrire une news</a></li>
@@ -49,16 +48,20 @@
             <li><a href="/admin/page/destroy/{{ $currentPage->id }}"><i class="fa fa-fw fa-trash"></i> Supprimer cette page</a></li>
             @endif
             <li role="separator" class="divider"></li>
-            <li><a href="/admin/training/index"><i class="fa fa-fw fa-graduation-cap"></i> Gérer les promotions (WIP)</a></li>
+            <li><a href="/admin/user/index"><i class="fa fa-fw fa-user"></i> Gérer les utilisateurs</a></li>
+            <li><a href="/admin/training/index"><i class="fa fa-fw fa-graduation-cap"></i> Gérer les promotions</a></li>
+            <li><a href="/admin/student/index"><i class="fa fa-fw fa-users"></i> Gérer les étudiants</a></li>
+            <li role="separator" class="divider"></li>
+            <li><a href="/admin/content/index"><i class="fa fa-fw fa-cog"></i> Paramètres globaux</a></li>
           </ul>
         </li>
         @if(isset($currentPage))
         <li><a href="/admin/page/edit/{{ $currentPage->id }}"><i class="fa fa-fw fa-pencil"></i></a></li>
-        <li><a href="/admin/page/destroy/{{ $currentPage->id }}"><i class="fa fa-trash"></i></a></li>
+        <li><a data-confirm="la page {{ $currentPage->title }}" href="/admin/page/destroy/{{ $currentPage->id }}"><i class="fa fa-trash"></i></a></li>
         @endif
         @if(isset($currentNews))
         <li><a href="/admin/news/edit/{{ $currentNews->id }}"><i class="fa fa-fw fa-pencil"></i></a></li>
-        <li><a href="/admin/news/destroy/{{ $currentNews->id }}"><i class="fa fa-fw fa-trash"></i></a></li>
+        <li><a data-confirm="la page {{ $currentNews->title }}" href="/admin/news/destroy/{{ $currentNews->id }}"><i class="fa fa-fw fa-trash"></i></a></li>
         @endif
         @endif
         @else
